cdk/requirements.txt
aws-cdk-lib>=2.100.0
constructs>=10.0.0
aws-cdk.aws-mwaa-alpha>=2.100.0a0
